test: updating serialization tests
