+++ pyproject.toml
@@ -197,7 +197,7 @@ exclude_lines = [
 
 [tool.coverage.run]
 source = ["src/ome_types", "src/ome_autogen"]
-
+omit = ["src/ome_types/_autogenerated/*"]
 
 # Entry points -- REMOVE ONCE XSDATA-PYDANTIC-BASEMODEL IS SEPARATE
 [project.entry-points."xsdata.plugins.class_types"]
