+++ src/ome_types/__init__.py
@@ -2,7 +2,7 @@ from importlib.metadata import PackageNotFoundError, version
 
 try:
     __version__ = version("ome-types")
-except PackageNotFoundError:
+except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 
 from ome_types import model
